Fix tearDown() methods in tests to properly tear down after exception
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@12355 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Tests/unitTests/dataspaces/VFSFactoryTest.java
@@ -38,8 +38,10 @@ public class VFSFactoryTest {
 
     @After
     public void tearDown() {
-        manager.close();
-        testFile.delete();
+        if (manager != null)
+            manager.close();
+        if (testFile != null)
+            testFile.delete();
     }
 
     @Test
